+++ src/ft_cache.cpp
@@ -2,13 +2,13 @@
 #include <cstdint>
 #include <cpp11/protect.hpp>
 
-FreetypeCache::FreetypeCache() 
+FreetypeCache::FreetypeCache()
   : error_code(0),
     glyphstore(),
     face_cache(16),
     size_cache(32),
     cur_id(),
-    cur_size(-1), 
+    cur_size(-1),
     cur_res(-1),
     cur_can_kern(false),
     cur_glyph(0)
@@ -24,48 +24,48 @@ FreetypeCache::~FreetypeCache() {
 
 bool FreetypeCache::load_font(const char* file, int index, double size, double res) {
   FaceID id(std::string(file), index);
-  
+
   if (current_face(id, size, res)) {
     return true;
   }
-  
+
   if (!load_face(id)) {
     return false;
   }
-  
+
   if (!load_size(id, size, res)) {
     return false;
   }
-  
+
   cur_id = id;
   cur_size = size;
   cur_res = res;
   glyphstore.clear();
-  
+
   cur_can_kern = FT_HAS_KERNING(face);
-  
+
   return true;
 }
 
 bool FreetypeCache::load_font(const char* file, int index) {
   std::string file_str(file);
   FaceID id(file_str, index);
-  
+
   if (id == cur_id) {
     return true;
   }
-  
+
   if (!load_face(id)) {
     return false;
   }
-  
+
   cur_id = id;
   cur_size = -1;
   cur_res = -1;
   glyphstore.clear();
-  
+
   cur_can_kern = FT_HAS_KERNING(face);
-  
+
   return true;
 }
 
@@ -73,7 +73,7 @@ bool FreetypeCache::load_face(FaceID face) {
   if (face == cur_id) {
     return true;
   }
-  
+
   FaceStore cached_face;
   if (face_cache.get(face, cached_face)) {
     this->face = cached_face.face;
@@ -115,7 +115,7 @@ bool FreetypeCache::load_size(FaceID face, double size, double res) {
   }
   FT_Size old_size = this->face->size;
   FT_Activate_Size(new_size);
-  
+
   if (cur_is_scalable) {
     err = FT_Set_Char_Size(this->face, 0, size * 64, res, res);
     if (err != 0) {
@@ -149,7 +149,7 @@ bool FreetypeCache::load_size(FaceID face, double size, double res) {
     if (!found_match && scaled_size >= largest_size) {
       best_match = largest_ind;
     }
-    
+
     err = FT_Select_Size(this->face, best_match);
     if (err != 0) {
       error_code = err;
@@ -163,9 +163,9 @@ bool FreetypeCache::load_size(FaceID face, double size, double res) {
       cached_face.sizes.erase(cached_id);
     }
   }
-  
+
   face_cache.add_size_id(face, id);
-  
+
   this->size = new_size;
   return true;
 }
@@ -175,13 +175,17 @@ bool FreetypeCache::has_glyph(uint32_t index) {
   return glyph_id != 0;
 }
 
-bool FreetypeCache::load_glyph(uint32_t index) {
+bool FreetypeCache::load_unicode(uint32_t index) {
   FT_UInt glyph_id = FT_Get_Char_Index(face, index);
+  return load_glyph(glyph_id);
+}
+
+bool FreetypeCache::load_glyph(FT_UInt id, int flags) {
   FT_Error err = 0;
-  err = FT_Load_Glyph(face, glyph_id, FT_LOAD_DEFAULT);
+  err = FT_Load_Glyph(face, id, flags);
   error_code = err;
   if (err == 0) {
-    cur_glyph = glyph_id;
+    cur_glyph = id;
   }
   return err == 0;
 }
@@ -205,9 +209,9 @@ FontInfo FreetypeCache::font_info() {
   res.n_sizes = face->num_fixed_sizes;
   res.n_charmaps = face->num_charmaps;
   res.bbox = {
-    FT_MulFix(face->bbox.xMin, size->metrics.x_scale), 
-    FT_MulFix(face->bbox.xMax, size->metrics.x_scale), 
-    FT_MulFix(face->bbox.yMin, size->metrics.y_scale), 
+    FT_MulFix(face->bbox.xMin, size->metrics.x_scale),
+    FT_MulFix(face->bbox.xMax, size->metrics.x_scale),
+    FT_MulFix(face->bbox.yMin, size->metrics.y_scale),
     FT_MulFix(face->bbox.yMax, size->metrics.y_scale)
   };
   res.max_ascend = FT_MulFix(face->ascender, size->metrics.y_scale);
@@ -217,19 +221,19 @@ FontInfo FreetypeCache::font_info() {
   res.lineheight = FT_MulFix(face->height, size->metrics.y_scale);
   res.underline_pos = FT_MulFix(face->underline_position, size->metrics.y_scale);
   res.underline_size = FT_MulFix(face->underline_thickness, size->metrics.y_scale);
-  
+
   return res;
 }
 
 GlyphInfo FreetypeCache::glyph_info() {
   GlyphInfo res = {};
-  
+
   res.index = cur_glyph;
   res.width = face->glyph->metrics.width;
   res.height = face->glyph->metrics.height;
   res.x_advance = face->glyph->advance.x;
   res.y_advance = face->glyph->advance.y;
-  
+
   if (res.y_advance != 0) { // Vertical
     res.x_bearing = face->glyph->metrics.vertBearingX;
     res.y_bearing = face->glyph->metrics.vertBearingY;
@@ -237,12 +241,12 @@ GlyphInfo FreetypeCache::glyph_info() {
     res.x_bearing = face->glyph->metrics.horiBearingX;
     res.y_bearing = face->glyph->metrics.horiBearingY;
   }
-  
+
   res.bbox = {res.x_bearing, res.x_bearing + res.width,
               res.y_bearing - res.height, res.y_bearing};
-  
+
   if (!cur_is_scalable) {
-    
+
     res.width *= unscaled_scaling;
     res.height *= unscaled_scaling;
     res.x_advance *= unscaled_scaling;
@@ -254,7 +258,7 @@ GlyphInfo FreetypeCache::glyph_info() {
     res.bbox[2] *= unscaled_scaling;
     res.bbox[3] *= unscaled_scaling;
   }
-  
+
   return res;
 }
 
@@ -262,9 +266,9 @@ GlyphInfo FreetypeCache::cached_glyph_info(uint32_t index, int& error) {
   std::map<uint32_t, GlyphInfo>::iterator cached_gi = glyphstore.find(index);
   GlyphInfo info = {};
   error = 0;
-  
+
   if (cached_gi == glyphstore.end()) {
-    if (load_glyph(index)) {
+    if (load_unicode(index)) {
       info = glyph_info();
       glyphstore[index] = info;
     } else {
@@ -273,7 +277,7 @@ GlyphInfo FreetypeCache::cached_glyph_info(uint32_t index, int& error) {
   } else {
     info = cached_gi->second;
   }
-  
+
   return info;
 }
 
@@ -291,33 +295,33 @@ bool FreetypeCache::get_kerning(uint32_t left, uint32_t right, long &x, long &y)
   y = 0;
   // Early exit
   if (!cur_can_kern) return true;
-  
+
   FT_UInt left_id = FT_Get_Char_Index(face, left);
   FT_UInt right_id = FT_Get_Char_Index(face, right);
-  
+
   FT_Vector delta = {};
-  
+
   FT_Error error = FT_Get_Kerning(face, left_id, right_id, FT_KERNING_DEFAULT, &delta);
-  
+
   if (error != 0) {
     error_code = error;
     return false;
   }
   x = delta.x;
   y = delta.y;
-  
+
   return true;
 }
 bool FreetypeCache::apply_kerning(uint32_t left, uint32_t right, long &x, long &y) {
   long delta_x = 0, delta_y = 0;
-  
+
   if (!get_kerning(left, right, delta_x, delta_y)) {
     return false;
   }
-  
+
   x += delta_x;
   y += delta_y;
-  
+
   return true;
 }